Make counting spectra less hacky
A non-managed model is added for storing the mapping from molecule to
spectra count. The molecule is backed by a database view which is
initialized from a custom sql query in a manually created migration
file. The molecule table uses this model instead of Molecule and can
therefore access the count as if it was an entry in the database.

The major benefit is the column being sortable and searchable now.
However, the column ordering had to be changed.